feat: 공통 함수로 분리
